Login almost finished / Link between XML file and setting screen
The plugin knows if the user of this project is already initialized (reads text between "password" tag) and shows the corresponding menu - login or registration. When registrated, generated token und encrypted password are saved in the xml file.
Before sending a request, the user types his name and mail address - these values are saved in the xml file as well. When they are changed, the xml gets updated. On new plugin start, name and mail fields are prefilled with values from xml.